Refactor CategoryDisplay component and extract useCategoryModal hook #83
