+++ client/app/ui/category/CategoryDisplay.tsx
@@ -1,9 +1,8 @@
-import React, { useState } from 'react'
-import { useRouter } from 'next/navigation'
+import React from 'react'
 import styled from 'styled-components'
 import { Category } from '@/app/types'
-import { deleteCategory, updateCategory } from '@/app/utils'
 import { CategoryDeleteModal, CategoryItem, CategoryUpdateModal } from '@/app/ui'
+import { useCategoryModal } from '@/app/utils'
 
 const CategoryDisplayContainer = styled.div`
   overflow-y: scroll;
@@ -22,46 +21,15 @@ interface Props {
 }
 
 export function CategoryDisplay({ categories }: Props) {
-  const router = useRouter()
-  const [isModalOpen, setIsModalOpen] = useState<'delete' | 'update' | undefined>()
-  const [targetCategory, setTargetCategory] = useState<Category | null>(null)
-
-  const openDeleteModal = (category: Category) => {
-    const component = document.getElementById(`${category.id}-hidden`)
-    if (component) {
-      component.style.minWidth = '0rem'
-      setIsModalOpen('delete')
-      setTargetCategory(category)
-    }
-  }
-
-  const openTargetModal = (category: Category) => {
-    const component = document.getElementById(`${category.id}-hidden`)
-    if (component) {
-      component.style.minWidth = '0rem'
-      setIsModalOpen('update')
-      setTargetCategory(category)
-    }
-  }
-
-  const closeModal = () => {
-    setIsModalOpen(undefined)
-    setTargetCategory(null)
-  }
-
-  const onClickDeleteButton = async () => {
-    if (!targetCategory) return
-    await deleteCategory(targetCategory.id)
-    closeModal()
-    router.refresh()
-  }
-
-  const onClickUpdateButton = async (title: string) => {
-    if (!targetCategory) return
-    await updateCategory(targetCategory.id, { title })
-    closeModal()
-    router.refresh()
-  }
+  const {
+    isModalOpen, //
+    targetCategory,
+    closeModal,
+    openDeleteModal,
+    openTargetModal,
+    onClickDeleteButton,
+    onClickUpdateButton
+  } = useCategoryModal()
 
   return (
     <CategoryDisplayContainer>
